refactor(stats): split statsService god object into layers (spec-80)
