--- src/modules/statistics-insights/interface-adapters/gateways/stats.gateway.ts
@@ -1,1 +1,0 @@
-export type { StatsGateway } from '@/modules/statistics-insights/entities/stats/stats.gateway.js';
