+++ src/tests/factories/persistedInsightsData.factory.ts
@@ -2,7 +2,7 @@ import type {
   PersistedDeveloperMetrics,
   PersistedInsightsData,
 } from '@/modules/statistics-insights/entities/insight/persistedInsightsData.js';
-import type { ReviewStats } from '@/modules/statistics-insights/services/statsService.js';
+import type { ReviewStats } from '@/modules/statistics-insights/entities/stats/projectStats.js';
 
 export class PersistedDeveloperMetricsFactory {
   static create(overrides: Partial<PersistedDeveloperMetrics> = {}): PersistedDeveloperMetrics {
